requirements.txt
pandas>=2.2,<3.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.2,<4.0
dbt-postgres>=1.10,<2.0
pytest>=8.0,<9.0
ruff>=0.6,<1.0
